+++ src/params/params.go
@@ -1,0 +1,46 @@
+package params
+
+/*
+CODE GENERATED AUTOMATICALLY WITH FIBER COIN CREATOR
+AVOID EDITING THIS MANUALLY
+*/
+
+const (
+	// Distribution locking parameteres
+
+	// MaxCoinSupply is the maximum supply of coins
+	MaxCoinSupply uint64 = 100000000
+	// DistributionAddressesTotal is the number of distribution addresses
+	DistributionAddressesTotal uint64 = 2
+	// DistributionAddressInitialBalance is the initial balance of each distribution address
+	DistributionAddressInitialBalance uint64 = MaxCoinSupply / DistributionAddressesTotal
+	// InitialUnlockedCount is the initial number of unlocked addresses
+	InitialUnlockedCount uint64 = 2
+	// UnlockAddressRate is the number of addresses to unlock per unlock time interval
+	UnlockAddressRate uint64 = 5
+	// UnlockTimeInterval is the distribution address unlock time interval, measured in seconds
+	// Once the InitialUnlockedCount is exhausted,
+	// UnlockAddressRate addresses will be unlocked per UnlockTimeInterval
+	UnlockTimeInterval uint64 = 31536000 // in seconds
+)
+
+var (
+	// Transaction verification parameters
+
+	// UserVerifyTxn transaction verification parameters for user-created transactions
+	UserVerifyTxn = VerifyTxn{
+		// BurnFactor can be overriden with `USER_BURN_FACTOR` env var
+		BurnFactor: 2,
+		// MaxTransactionSize can be overriden with `USER_MAX_TXN_SIZE` env var
+		MaxTransactionSize: 32768, // in bytes
+		// MaxDropletPrecision can be overriden with `USER_MAX_DECIMALS` env var
+		MaxDropletPrecision: 3,
+	}
+)
+
+// distributionAddresses are addresses that received coins from the genesis address in the first block,
+// used to calculate current and max supply and do distribution timelocking
+var distributionAddresses = [DistributionAddressesTotal]string{
+	"TkyD4wD64UE6M5BkNQA17zaf7Xcg4AufwX",
+	"2PBcLADETphmqWV7sujRZdh3UcabssgKAEB",
+}