+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.6.3 (NOT YET RELEASED)
 * Additions
   * Prototype for CX chains (CX programs stored on the Skycoin blockchain)
+  * Added/forked the newcoin and skycoin-cli commands to the CX repository.
 * Libraries
 * Fixed issues
   * #292: Compilation error when left hand side of an assignment expression is a struct field.